Add files via upload
Write a program to calculate area of shapes. Your program should be capable of calculating the area of a square, rectangle, triangle and a circle. The user should be presented with options to select the shape. Based on which shape is chooses by the user, the program should ask for the appropriate input and print the resulting area on the screen